Added sweet alert

+++ src/Components/CustomerComponent/UpdateCustomer/Update.js
@@ -20,7 +20,7 @@ class App extends Component {
     };
 
     didMountTick = () => {
-        this.props.start(true);
+        // this.props.start(true);
         axios.get(`customers`)
             .then(response => {
                 console.log(response.data);
@@ -67,8 +67,9 @@ class App extends Component {
             this.noAddressBackground();
         }
 
-        if (this.state.name.trim() !== '' && this.state.address.trim() !== '') {
-            let id = document.getElementById("example1").value;
+        let id = document.getElementById("example1").value;
+
+        if (this.state.name.trim() !== '' && this.state.address.trim() !== ''  && id !== '') {
 
             const customerObj={
                 id:id,
@@ -98,7 +99,7 @@ class App extends Component {
 
         }else{
             swal({
-                text: "Please fill all textfield...!",
+                text: "Please fill all field...!",
                 icon: "warning",
                 button: "Okay!",
             });
@@ -106,24 +107,33 @@ class App extends Component {
     };
 
     removesubmit = () => {
-        this.props.start(true);
+        // this.props.start(true);
         let id = document.getElementById("example1").value;
         console.log(id);
+        if(id!==""){
+            axios.delete(`customers/` + id)
+                .then(response1 => {
+                    swal({
+                        text: "Removed!",
+                        icon: "success",
+                        button: "Okay!",
+                    });
 
-        axios.delete(`customers/` + id)
-            .then(response1 => {
-                swal({
-                    text: "Removed!",
-                    icon: "success",
-                    button: "Okay!",
+                    this.didMountTick()
+                })
+
+                .catch(error => {
+                    console.log("error: " + error)
                 });
-
-                this.didMountTick()
-            })
-
-            .catch(error => {
-                console.log("error: " + error)
+        }else{
+            swal({
+                text: "Please select id...!",
+                icon: "warning",
+                button: "Okay!",
             });
+        }
+
+
     };
 
 
@@ -167,9 +177,7 @@ class App extends Component {
                     <label htmlFor="example1">Select Id</label>
                     <select onChange={(event) => this.idChange(event)} className="form-control" id="example1"
                             style={{
-                                width: '30%',
-                                borderRadius: '20px',
-                                background: 'linear-gradient(to right, rgb(219, 230, 246), rgb(197, 121, 109))'
+                                width: '30%',height:'34px',background:'linear-gradient(to left, antiquewhite, white)'
                             }}>
                         {options}
